Refactor code and folder structure
