+++ backend/cmd/api-server/main.go
@@ -1,0 +1,24 @@
+package main
+
+import (
+	"log"
+
+	"github.com/gofiber/fiber/v2"
+	"github.com/gofiber/fiber/v2/middleware/cors"
+	"github.com/thepokenik/elf-note/internal/users"
+)
+
+func main() {
+
+	app := fiber.New()
+
+	app.Use(cors.New(cors.Config{
+		AllowOrigins: "http://localhost:3000",
+		AllowHeaders: "Origin, Content-Type, Accept",
+	}))
+
+	users.SetRoutes(app)
+
+	log.Fatal(app.Listen(":4000"))
+
+}
